ADD 피드 CatTags, Tag 수정, 유효성 검증 추가 #78

+++ back-end/neighbor_cats/src/main/java/org/dnd3/udongsa/neighborcats/feed/service/FeedServiceImpl.java
@@ -7,9 +7,7 @@ import java.util.Objects;
 
 import org.dnd3.udongsa.neighborcats.address.Address;
 import org.dnd3.udongsa.neighborcats.cat.dto.CatDto;
-import org.dnd3.udongsa.neighborcats.cat.entity.Cat;
-import org.dnd3.udongsa.neighborcats.cat.entity.CatMapper;
-import org.dnd3.udongsa.neighborcats.cat.service.CatService;
+import org.dnd3.udongsa.neighborcats.exception.CustomException;
 import org.dnd3.udongsa.neighborcats.feed.dto.FeedCommentDto;
 import org.dnd3.udongsa.neighborcats.feed.dto.FeedDto;
 import org.dnd3.udongsa.neighborcats.feed.dto.FeedModifyDto;
@@ -19,9 +17,12 @@ import org.dnd3.udongsa.neighborcats.feed.dto.PagingDto;
 import org.dnd3.udongsa.neighborcats.feed.entity.EFilterType;
 import org.dnd3.udongsa.neighborcats.feed.entity.ESortType;
 import org.dnd3.udongsa.neighborcats.feed.entity.Feed;
+import org.dnd3.udongsa.neighborcats.feed.entity.FeedImg;
 import org.dnd3.udongsa.neighborcats.feed.entity.FeedMapper;
 import org.dnd3.udongsa.neighborcats.feed.repository.FeedRepository;
 import org.dnd3.udongsa.neighborcats.imgfile.dto.ImgFileDto;
+import org.dnd3.udongsa.neighborcats.role.ERole;
+import org.dnd3.udongsa.neighborcats.role.Role;
 import org.dnd3.udongsa.neighborcats.security.service.SecurityContextService;
 import org.dnd3.udongsa.neighborcats.servant.dto.AuthorDto;
 import org.dnd3.udongsa.neighborcats.servant.entity.Servant;
@@ -34,8 +35,10 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort.Direction;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
 import lombok.RequiredArgsConstructor;
 
@@ -51,7 +54,7 @@ public class FeedServiceImpl implements FeedService {
   private final FeedLikeService feedLikeService;
   private final TimeDescService timeDescService;
   private final ServantService servantService;
-  private final CatService catService;
+  private final FeedCatService feedCatService;
 
   @Override
   @Transactional(readOnly = true)
@@ -78,7 +81,7 @@ public class FeedServiceImpl implements FeedService {
   }
 
   private FeedDto toDto(Feed feed){
-    List<TagDto> feedTags = feedTagService.getAllByFeed(feed);
+    TagDto tagDto = feedTagService.findTagDtoByFeed(feed);
     List<FeedCommentDto> comments = commentService.getAllByFeed(feed);
     List<ImgFileDto> imgDtos = feedImgService.getAllByFeed(feed);
     AuthorDto authorDto = ServantMapper.map(feed.getAuthor());
@@ -87,8 +90,8 @@ public class FeedServiceImpl implements FeedService {
     int numberOfComments = comments.size();
     LocalDateTime createdDateTime = feed.getCreatedAt();
     String timeDesc = timeDescService.generate(createdDateTime);
-    CatDto catDto = CatMapper.map(feed.getCat());
-    FeedDto feedDto = FeedMapper.map(feed, feedTags, comments, imgDtos, authorDto, isLike, numberOfLikes, numberOfComments, createdDateTime, timeDesc, catDto);
+    List<CatDto> catDtos = feedCatService.findAllByFeed(feed);
+    FeedDto feedDto = FeedMapper.map(feed, tagDto, comments, imgDtos, authorDto, isLike, numberOfLikes, numberOfComments, createdDateTime, timeDesc, catDtos);
     return feedDto;
   }
 
@@ -96,12 +99,29 @@ public class FeedServiceImpl implements FeedService {
   @Transactional
   public FeedDto save(FeedSaveDto saveDto) {
     Servant author = servantService.findServantByEmail(securityService.getLoggedUserEmail());
-    Cat cat = catService.findCatById(saveDto.getCatId());
-    Feed feed = FeedMapper.map(saveDto, author, cat);
+    Feed feed = FeedMapper.map(saveDto, author);
     repo.save(feed);
+    if(Objects.isNull(saveDto.getImgFiles()) || saveDto.getImgFiles().size() == 0){
+      throw new CustomException(HttpStatus.BAD_REQUEST, "이미지를 한 개 이상 업로드해야 합니다.");
+    }
     feedImgService.save(saveDto.getImgFiles(), feed);
-    feedTagService.save(saveDto.getTagIds(), feed);
+    if(Objects.nonNull(saveDto.getTagId())){
+      feedTagService.save(saveDto.getTagId(), feed);
+    }
+    if(Objects.nonNull(saveDto.getCatIds()) && saveDto.getCatIds().size() > 0){
+      validateCat(saveDto.getCatIds());
+      feedCatService.save(saveDto.getCatIds(), feed);
+    }
     return toDto(feed);
+  }
+
+  private void validateCat(List<Long> catIds) {
+    String email = securityService.getLoggedUserEmail();
+    Servant servant = servantService.findServantByEmail(email);
+    boolean doesHave = feedCatService.doesHaveCat(servant, catIds);
+    if(!doesHave){
+      throw new CustomException(HttpStatus.BAD_REQUEST, "해당 고양이를 가지고 있지 않습니다.", "cats: {}, servantEmail: {}", catIds, email);
+    }
   }
 
   @Override
@@ -115,25 +135,53 @@ public class FeedServiceImpl implements FeedService {
   @Transactional
   public FeedDto delete(Long id) {
     Feed feed = repo.findById(id).orElseThrow();
+    validateAuthor(feed.getAuthor());
     feedTagService.deleteByFeed(feed);
     commentService.deleteByFeed(feed);
     feedImgService.deleteByFeed(feed);
     feedLikeService.deleteByFeed(feed);
+    feedCatService.deleteByFeed(feed);
     FeedDto feedDto = new FeedDto();
     feedDto.setId(feed.getId());
     repo.delete(feed);
     return feedDto;
   }
 
+  private void validateAuthor(Servant author) {
+    String email = securityService.getLoggedUserEmail();
+    Servant servant = servantService.findServantByEmail(email);
+    if(servant.getId() == author.getId()) return;
+    for(Role role : author.getRoles()){
+      if(role.getName() == ERole.ROLE_ADMIN) return;
+    }
+    throw new CustomException(HttpStatus.FORBIDDEN, "작성/수정 권한이 없습니다.", "작성자 이메일: {}, 요청자 이메일: {}", author.getEmail(), servant.getEmail());      
+  }
+
   @Override
   @Transactional
   public FeedDto modify(Long id, FeedModifyDto modifyDto) {
     Feed persist = repo.findById(id).orElseThrow();   
-    Cat modifyCat = catService.findCatById(modifyDto.getCatId());
-    persist.update(modifyDto.getContent(), modifyCat);
-    feedImgService.deleteByImgFileIds(persist, modifyDto.getRemoveImgFileIds());
-    feedImgService.save(modifyDto.getInsertImgFiles(), persist);
-    feedTagService.update(persist, modifyDto.getFeedTagIds());
+    validateAuthor(persist.getAuthor());
+    persist.update(modifyDto.getContent());
+    List<Long> removedImgIds = modifyDto.getRemoveImgFileIds();
+    List<MultipartFile> insertImgFiles = modifyDto.getInsertImgFiles();
+    if(Objects.nonNull(removedImgIds) && removedImgIds.size() > 0){
+      feedImgService.deleteByImgFileIds(persist, modifyDto.getRemoveImgFileIds());
+    }
+    if(Objects.nonNull(insertImgFiles) && insertImgFiles.size() > 0){
+      feedImgService.save(modifyDto.getInsertImgFiles(), persist);
+    }
+    if(Objects.nonNull(modifyDto.getTagId())){
+      feedTagService.update(persist, modifyDto.getTagId());
+    }
+    if(Objects.nonNull(modifyDto.getCatIds())){
+      validateCat(modifyDto.getCatIds());
+      feedCatService.update(modifyDto.getCatIds(), persist);
+    }
+    List<ImgFileDto> imgFileDtos = feedImgService.getAllByFeed(persist);
+    if(imgFileDtos.size() == 0){
+      throw new CustomException(HttpStatus.BAD_REQUEST, "이미지를 한 개 이상 업로드해야 합니다.");
+    }
     return toDto(persist);
   }
